+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+`1.1.0`
+- Default mouse wheel buttons (emotes, shapeshifts) will temporarily override mouse lock in first person/action mode while pressed
+- Added keybind for completing journal quests (minus default)
+
 `1.0.0`
-- Thunderstore release
+- Initial release